feat: start refactoring
