+++ src/scss/mail-sender.scss
@@ -274,5 +274,9 @@
       padding-left: 6px;
       color: #ff6666;
     }
+    &:hover {
+      cursor: pointer;
+    }
   }
+
 }
