+++ src/utils/form.js
@@ -5,3 +5,9 @@ export const MIN_INPUT_LENGTH = 5;
 export const isTotalAttachmentsSizeExceeded = (totalAttachmentsSize) => {
   return totalAttachmentsSize > ATTACHMENTS_SIZE_LIMIT ? true : false;
 };
+
+export const truncateString = (value, num) => {
+  return value.length <= num
+    ? value
+    : value.slice(0, 19) + '...' + value.slice(value.indexOf('.'), value.length);
+};
